+++ src/app/modules/challenge/components/challenge-form/challenge-form.component.ts
@@ -44,6 +44,7 @@ export class ChallengeFormComponent implements AfterViewInit, OnDestroy {
   }
 
   languages: Language[] = []
+  selectedLanguageId: string = ''
   selectedTags: string[] = []
   currentTags: any[] = []
 
@@ -76,7 +77,6 @@ export class ChallengeFormComponent implements AfterViewInit, OnDestroy {
     @Inject(TranslateService) readonly translate: TranslateService
   ) {
     this.loadLanguages()
-    this.loadTags()
     translate.addLangs(['en', 'es', 'ca'])
     translate.setDefaultLang('es')
     translate.use('es')
@@ -158,6 +158,9 @@ export class ChallengeFormComponent implements AfterViewInit, OnDestroy {
 
   onLanguageChange (language: string): void {
     this.challenge.language = language
+    const selectedLang = this.languages.find(lang => lang.language_name === language)
+    this.selectedLanguageId = (selectedLang != null) ? selectedLang.id_language : ''
+    this.loadTags()
     /* istanbul ignore next */
     // Actualiza CodeMirror con el nuevo lenguaje
     if (this.editor != null) {
@@ -178,14 +181,16 @@ export class ChallengeFormComponent implements AfterViewInit, OnDestroy {
   }
 
   loadTags (): void {
-    this.challengeFormService.getTags().subscribe({
+    if (this.selectedLanguageId.length === 0) {
+      return
+    }
+    this.challengeFormService.getTagsByLanguage(this.selectedLanguageId).subscribe({
       next: (response: TagResponse) => {
         this.currentTags = response.results ?? []
-
         this.selectedTags = []
       },
       error: (error) => {
-        console.error('Error al obtener las etiquetas:', error)
+        console.error('Error fetching tags:', error)
         this.currentTags = []
         this.selectedTags = []
       }
@@ -200,7 +205,6 @@ export class ChallengeFormComponent implements AfterViewInit, OnDestroy {
     return (
       this.challenge.challengeTitle.trim() !== '' &&
       this.challenge.description.trim() !== '' &&
-      isLanguageValid &&
       isLanguageValid &&
       this.challenge.solution.trim() !== ''
     )