+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to
 
 - Configuration for SonarCloud compatibility: added sonar-project.properties with coverage paths and exclusions (#493)
 - Fixed missing braces in en.json and es.json (#492)
+- Refactored challenge-form service and component to load tags based on selected language (#470)
 - Visual refactor of icons in challenge-card (rocket icon, hover effects, tooltip, and layout reorder) (#471)
 
 ### [ita-challenges-frontend-3.1.35-RELEASE] - 2025-05-28